+++ pages/home/urls.py
@@ -1,8 +1,7 @@
 from django.urls import path
 
-from .views import home_map_settings, wdqms_reports
+from .views import home_map_settings
 
 urlpatterns = [
     path("api/home-map/settings", home_map_settings, name="home-map-settings"),
-    path("wdqms-reports/", wdqms_reports, name="wdqms-reports"),
 ]
