+++ pages/home/views.py
@@ -3,12 +3,8 @@ from django.http import JsonResponse
 from django.urls import reverse
 from wagtail.api.v2.utils import get_full_url
 from django.shortcuts import render
-import json
 from base.models import OrganisationSetting
 from pages.home.models import HomeMapSettings
-from climweb_wdqms.models import Station,Transmission
-from django.db.models.functions import TruncMonth
-from django.db.models import Max
 
 def home_map_settings(request):
     config = {
@@ -44,29 +40,3 @@ def home_map_settings(request):
         })
 
     return JsonResponse(config)
-
-
-def wdqms_reports(request):
-
-    stations = Station.objects.all()
-    transmissions = Transmission.objects.all()
-    variables = []
-    latest_date = None
-    years = []
-
-    if transmissions:
-        variables = transmissions.values_list('variable', flat=True).distinct()
-        years = transmissions.values_list('received_date__year', flat=True).distinct()
-        latest_date = transmissions.filter(variable=variables[0]).order_by('received_date').values_list('received_date__date',  flat=True).last()
-
-    abm_settings = AdminBoundarySettings.for_request(request)
-
-    abm_extents = abm_settings.combined_countries_bounds
-
-    return render(request, "home/wdqms_report/report_index.html", {
-        'stations': stations,
-        'variables':variables,
-        'latest_date':latest_date,
-        'years':years,
-        'bounds':abm_extents,
-    })
