+++ requirements.txt
@@ -22,7 +22,7 @@ click-didyoumean==0.3.0
 click-plugins==1.1.1
 click-repl==0.2.0
 cligj==0.7.2
-climweb-wdqms==0.0.4
+climweb-wdqms==0.0.5
 contourpy==1.0.7
 cryptography==41.0.4
 cycler==0.11.0
